+++ src/calc-size-robust.h
@@ -1,0 +1,2 @@
+
+int calc_size_robust(SEXP robj);